services/es_sync/requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
elasticsearch>=8.10.0,<9.0.0

